backend/requirements.txt
fastapi
uvicorn
python-jose
passlib
minio
requests
proxmoxer
pyotp
